Parser: move shared YAML partials to dedicated files
